Access internal services via public IP
It does not seem that NATS can be accessed via 127.0.0.1 if it
binds itself to one of the network interfaces; so the
jobs need to access NATS via the public IP (until internal
DNS is supported).

Therefore all access to all things via public IP.

Need to ensure all the ports are open publicly.