cf-release branches cloned into separate paths; current branch in system_config [fixes #95]
Branches are cleanly kept in separate folders since
I cannot trust git to cleanly switch branches, pull down
the matching branch from the remote, and not do silly
merging of the previous branch.

Defaulting to staging branch for the moment until
the required patches in staging are merged into master
and ultimately into a final release.